requirements.txt
sentence-transformers>=2.7.0
torch>=2.0.0
transformers>=4.40.0
umap-learn>=0.5.6
hdbscan>=0.8.33
faiss-cpu>=1.8.0
plotly>=5.20.0
pandas>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
numpy>=1.26.0
